+++ pymongo/__init__.py
@@ -86,7 +86,8 @@ from pymongo.common import (MIN_SUPPORTED_WIRE_VERSION,
 from pymongo.cursor import CursorType
 from pymongo.mongo_client import MongoClient
 from pymongo.mongo_replica_set_client import MongoReplicaSetClient
-from pymongo.operations import (InsertOne,
+from pymongo.operations import (IndexModel,
+                                InsertOne,
                                 DeleteOne,
                                 DeleteMany,
                                 UpdateOne,