+++ pymongo/operations.py
@@ -15,6 +15,7 @@
 """Operation class definitions."""
 
 from pymongo.common import validate_boolean, validate_is_mapping
+from pymongo.helpers import _gen_index_name, _index_document, _index_list
 
 
 class _WriteOp(object):
@@ -173,3 +174,65 @@ class UpdateMany(_WriteOp):
                                            self._doc,
                                            self._upsert)
 
+
+class IndexModel(object):
+    """Represents an index to create."""
+
+    __slots__ = ("__document",)
+
+    def __init__(self, keys, **kwargs):
+        """Create an Index instance.
+
+        For use with :meth:`~pymongo.collection.Collection.create_indexes`.
+
+        Takes either a single key or a list of (key, direction) pairs.
+        The key(s) must be an instance of :class:`basestring`
+        (:class:`str` in python 3), and the direction(s) must be one of
+        (:data:`~pymongo.ASCENDING`, :data:`~pymongo.DESCENDING`,
+        :data:`~pymongo.GEO2D`, :data:`~pymongo.GEOHAYSTACK`,
+        :data:`~pymongo.GEOSPHERE`, :data:`~pymongo.HASHED`,
+        :data:`~pymongo.TEXT`).
+
+        Valid options include, but are not limited to:
+
+          - `name`: custom name to use for this index - if none is
+            given, a name will be generated.
+          - `unique`: if ``True`` creates a uniqueness constraint on the index.
+          - `background`: if ``True`` this index should be created in the
+            background.
+          - `sparse`: if ``True``, omit from the index any documents that lack
+            the indexed field.
+          - `bucketSize`: for use with geoHaystack indexes.
+            Number of documents to group together within a certain proximity
+            to a given longitude and latitude.
+          - `min`: minimum value for keys in a :data:`~pymongo.GEO2D`
+            index.
+          - `max`: maximum value for keys in a :data:`~pymongo.GEO2D`
+            index.
+          - `expireAfterSeconds`: <int> Used to create an expiring (TTL)
+            collection. MongoDB will automatically delete documents from
+            this collection after <int> seconds. The indexed field must
+            be a UTC datetime or the data will not expire.
+
+        See the MongoDB documentation for a full list of supported options by
+        server version.
+
+        :Parameters:
+          - `keys`: a single key or a list of (key, direction)
+            pairs specifying the index to create
+          - `**kwargs` (optional): any additional index creation
+            options (see the above list) should be passed as keyword
+            arguments
+        """
+        keys = _index_list(keys)
+        if "name" not in kwargs:
+            kwargs["name"] = _gen_index_name(keys)
+        kwargs["key"] = _index_document(keys)
+        self.__document = kwargs
+
+    @property
+    def document(self):
+        """An index document suitable for passing to the createIndexes
+        command.
+        """
+        return self.__document