+++ doc/changelog.rst
@@ -267,6 +267,7 @@ The following methods have been added:
 - :meth:`~pymongo.collection.Collection.find_one_and_replace`
 - :meth:`~pymongo.collection.Collection.find_one_and_update`
 - :meth:`~pymongo.collection.Collection.with_options`
+- :meth:`~pymongo.collection.Collection.create_indexes`
 
 The following methods have changed: